ai-backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# LangChain
langchain==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.0
langchain-community==0.3.14
langgraph==0.2.60

# AI/ML
openai==1.58.1
anthropic==0.42.0
tiktoken==0.8.0

# Database
supabase==2.10.0

# Tools
tavily-python==0.5.0
duckduckgo-search==7.2.1

# Utils
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx>=0.23.0,<0.28
aiohttp==3.11.11

# Vector Store
chromadb==0.5.23
sentence-transformers==3.3.1
